compose: Add hint text to topic typeahead dropdown.
Fixes #23295

With the autocomplete dropdown, some users face trouble or uncertainty while
selecting topics. The autocomplete dropdown may encourage users to find and
select an existing topic rather than starting a new topic. The changes adds
a hint text to the topic typeahead dropdown to encourage users to either
start a new topic or select an existing topic. The changes also adds a new
indicator to the topic typeahead dropdown while typing to the topic box.
This ensures users that they are on track and they are doing the right
things.